+++ src/components/Form/VDatePicker.vue
@@ -1,12 +1,29 @@
 <template>
-  <input v-if="date === ''" type="date" :value=currentDate>
-  <input v-else type="date" :value=date>
+  <div class="date-picker">
+    <label v-if=title class="font-normal">{{title}}</label>
+    <span v-if=subtitle class="info">{{subtitle}}</span>
+    <input v-if="date === ''" type="date" :value=currentDate>
+    <input v-else type="date" :value=date>
+    <span v-if=information class="info text-center">{{information}}</span>
+  </div>
 </template>
 
 <script>
 export default {
   props: {
     date: {
+      type: String,
+      default: '',
+    },
+    title: {
+      type: String,
+      default: '',
+    },
+    subtitle: {
+      type: String,
+      default: '',
+    },
+    information: {
       type: String,
       default: '',
     },